Adds HIP0019 readme, consolidates on main.debug()
